requirements.txt
tensorflow==2.13.0
opencv-python==4.8.1.78
mediapipe==0.10.7
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
scikit-learn==1.3.0
pyttsx3==2.90
python-multipart==0.0.6
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.3
seaborn==0.13.0
matplotlib==3.8.1
pillow==10.1.0